+++ src/client/__init__.py
@@ -1,13 +1,19 @@
 from pssapi.entities import Alliance as PssAlliance
 from pssapi.entities import User as PssUser
 
-from .model import Collection, CollectionMetadata, PssFleetDataClient, exceptions
+from . import core, models
+from .client import ClientConfig, PssFleetDataClient
+from .core import exceptions
+from .models import Collection, CollectionMetadata
 
 
 __all__ = [
     # Modules
+    core.__name__,
     exceptions.__name__,
+    models.__name__,
     # Classes
+    ClientConfig.__name__,
     Collection.__name__,
     CollectionMetadata.__name__,
     PssAlliance.__name__,
