--- src/client/model/__init__.py
@@ -1,19 +1,0 @@
-from . import api, client, converters, enums, exceptions
-from .client import PssFleetDataClient
-from .models import AllianceHistory, Collection, CollectionMetadata, UserHistory
-
-
-__all__ = [
-    # modules
-    api.__name__,
-    client.__name__,
-    converters.__name__,
-    enums.__name__,
-    exceptions.__name__,
-    # classes
-    AllianceHistory.__name__,
-    Collection.__name__,
-    CollectionMetadata.__name__,
-    PssFleetDataClient.__name__,
-    UserHistory.__name__,
-]
