+++ src/client/models/__init__.py
@@ -1,0 +1,14 @@
+from . import api_models, converters
+from .client_models import AllianceHistory, Collection, CollectionMetadata, UserHistory
+
+
+__all__ = [
+    # modules
+    api_models.__name__,
+    converters.__name__,
+    # classes
+    AllianceHistory.__name__,
+    Collection.__name__,
+    CollectionMetadata.__name__,
+    UserHistory.__name__,
+]
