+++ src/client/models/api_models.py
@@ -3,9 +3,9 @@ from typing import Annotated, Any, Optional, Union
 
 from pydantic import BaseModel, Field, field_validator
 
-from .. import utils
-from ..config import CONFIG
-from .enums import UserAllianceMembershipEncoded
+from ..core import utils
+from ..core.config import CONFIG
+from ..core.enums import UserAllianceMembershipEncoded
 
 
 DATETIME = Annotated[datetime, Field(ge=CONFIG.pss_start_date)]
