+++ tests/10_utils/test_add_timezone_utc.py
@@ -3,7 +3,7 @@ from datetime import datetime
 import pytest
 from dateutil.parser import parse as parse_datetime
 
-from client.utils import add_timezone_utc
+from client.core.utils import add_timezone_utc
 
 
 test_cases_invalid = [
