+++ tests/30_routes/test_get_collection.py
@@ -3,8 +3,8 @@ from typing import Callable
 import pytest
 
 from client import PssFleetDataClient
-from client.model import Collection
-from client.model.exceptions import CollectionNotFoundError, InvalidCollectionIdError
+from client.core.exceptions import CollectionNotFoundError, InvalidCollectionIdError
+from client.models import Collection
 
 
 @pytest.mark.usefixtures("mock_response_collections_collectionId_get_200")
