+++ tests/30_routes/test_get_top_100_users_from_collection.py
@@ -4,8 +4,8 @@ import pytest
 from pssapi.entities import User as PssUser
 
 from client import PssFleetDataClient
-from client.model import Collection
-from client.model.exceptions import CollectionNotFoundError, InvalidCollectionIdError
+from client.core.exceptions import CollectionNotFoundError, InvalidCollectionIdError
+from client.models import Collection
 
 
 @pytest.mark.usefixtures("mock_response_collections_collectionId_users_get_200")
