+++ tests/fixtures/client_objects.py
@@ -4,8 +4,8 @@ import pytest
 from pssapi.entities import Alliance as PssAlliance
 from pssapi.entities import User as PssUser
 
-from client import utils
-from client.model import AllianceHistory, Collection, CollectionMetadata, UserHistory
+from client.core import utils
+from client.models import AllianceHistory, Collection, CollectionMetadata, UserHistory
 
 
 # Client objects
